% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetareg.R
\docType{package}
\name{abetareg}
\alias{abetareg}
\title{abetareg: Log-likelihood Adjustment for Beta Regression Models}
\description{
This package preforms log-likelihood adjustment for beta
regression, using maximum likelihood estimation, by 'betareg' package
\href{https://cran.r-project.org/package=betareg}{betareg}.
}
