% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_object.R
\name{lax-internal}
\alias{lax-internal}
\alias{adj_object}
\title{Internal lax functions}
\usage{
adj_object(x, cluster = NULL, use_vcov = TRUE, ...)
}
\description{
Internal lax functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
