% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikVec.R
\name{logLikVec}
\alias{logLikVec}
\alias{logLikVec.betareg}
\title{Evaluate Log-likelihood Contributions From Specific Observations}
\usage{
logLikVec(object, ...)

\method{logLikVec}{betareg}(object, pars = NULL, ...)
}
\arguments{
\item{object}{A fitted model object.(do we need to specify betareg here?)}

\item{...}{Further arguments.}

\item{pars}{A vector of coefficients.}
}
\description{
This is a generic function for calculating log-likelihood contributions from
individual observations for a fitted model.
}
