% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.logLikVec}
\alias{logLik.logLikVec}
\title{Sum Log-likelihood Contributions From Individual Observations}
\usage{
\method{logLik}{logLikVec}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"logLikVec"} return from a
\code{logLikVec} method.}

\item{...}{Further arguments.}
}
\description{
S3 logLik method for logLikVec objects
}
