% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikVec.R
\name{logLikVec}
\alias{logLikVec}
\alias{logLikVec.betareg}
\title{Evaluate Log-likelihood Contributions From Specific Observations}
\usage{
logLikVec(object, ...)

\method{logLikVec}{betareg}(object, pars = NULL, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments.}

\item{pars}{A vector of coefficients.}
}
\value{
A vector of individual log-likelihood contributions of class
\code{"logLikVec"}, computed from \code{"betareg"} object,
with two additional attributes:
\item{nobs}{number of observations}
\item{df}{degrees of freedom}
}
\description{
This is a generic function for calculating log-likelihood contributions from
individual observations for a fitted model.
}
\examples{
data("sp", package = "abetareg")

## Data scaling suggested by Smithson and Verkuilen (2006)
x <- sp$performance
x1 <- (x - 10) / (100 - 10)
x2 <- (x1 * (length(x) - 1) + 0.5) / length(x)
sp$performance <- x2

## Fit the betareg model
library(betareg)
fit <- betareg(performance ~ studyh + ea + previous | studyh + sleeph, data = sp)
summary(fit)

## Check if the sum of individual log-likelihood == total log-likelihood
all.equal(fit$loglik, sum(logLikVec(fit)), check.attributes = FALSE)

}
\references{
Smithson, Michael & Verkuilen, Jay. (2006). A better lemon
squeezer? Maximum-likelihood regression with beta-distributed
dependent variables. Psychological methods. 11. 54-71.
\doi{10.1037/1082-989X.11.1.54}
}
